% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputOccDataSummary.R
\name{InputOccDataSummary}
\alias{InputOccDataSummary}
\title{InputOccDataSummary}
\usage{
InputOccDataSummary(indata = "data/model_runs/", output_path = "output/",
  minNyear = 2, minNrecs = 50, save_name = "cirrus",
  outhwaite_spp_clean = TRUE)
}
\arguments{
\item{indata}{The file path to a folder contain .rdata output files 
from the Data_cleaning.function}

\item{output_path}{The file path for the location where we want to 
save the outputs}

\item{minNyear}{The minimum number of years of records a site needs to be 
included in the final modelled dataset}

\item{minNrecs}{The total (across all regions) minimum number of records a 
species needs to be included in the final modelled dataset}

\item{save_name}{the name of the file to be saved}

\item{outhwaite_spp_clean}{Set to TRUE or FALSE. ONly include species included 
in Outhwaite's data paper}
}
\value{
A .csv file summarising the number of species included for each
region from the cleaned dataset
}
\description{
THIS FUNCTION IS STILL UNDER REVIEW - This function takes .rdata 
files created by the Data_cleaning.function and calculates how many species 
have records in each region follwoing the data cleaning process.
}
