% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotOcc.R
\name{PlotOcc}
\alias{PlotOcc}
\title{PlotOcc - Produce time-series plots of annual occupancy estimates}
\usage{
PlotOcc(
  indata = "../data/model_runs/",
  output_path = "../output/",
  REGION_IN_Q = "^psi.fs\\\\[",
  y_axis_choice = "variable",
  cluster_run = "jasmin",
  jasmin_min_year = 1970
)
}
\arguments{
\item{indata}{The folder containing the species-level occupancy model outputs}

\item{output_path}{The location where the output plots will be saved.}

\item{REGION_IN_Q}{The region for which the time-series that will be plotted.
UK based occupancy model examples include,
 'psi.fs', 'psi.fs.r_GB', 'psi.fs.r_ENGLAND'.}

\item{y_axis_choice}{Two choices here, "variable" or "fixed". Fixed, fixes the 
y axis limits between 0 and 1. Variable sets the y axis limits equal to the 
maximum and minimum plotted values.}

\item{cluster_run}{This is a UK specific parameter. "Jasmin" is used for outputs
from Jasmin occupancy model runs.  "Cirrus" is used for standard .rdata occupancy 
model outputs, such as those from the occDetFunc function from the R package Sparta.}

\item{jasmin_min_year}{If running on Jasmin outputs, the first year must be specified}
}
\value{
A .png file for each species containing time-series plots of annual occupancy.
}
\description{
This function plots species-level annual occupancy time-series. The
function plots all species within the input folder.  Currently had a hard coded name_
match file that I will needs to remove
}
