% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RhatSummary.R
\name{RhatSummary}
\alias{RhatSummary}
\title{RhatSummary - Rhat Summary function}
\usage{
RhatSummary(
  indata = "../data/model_runs/",
  output_path = "../rhat_summary/",
  REGION_IN_Q = "psi.fs",
  group_run = "article_17"
)
}
\arguments{
\item{indata}{The file path to a location containing .rdata files
of the bugs outputs, one for each species}

\item{output_path}{The location where the output will be saved.}

\item{REGION_IN_Q}{The region for which the rhat values will be 
extracted. UK based occupancy model examples include,
 'psi.fs', 'psi.fs.r_GB', 'psi.fs.r_ENLGAND'.}

\item{group_run}{The name of the group of species that were run.
This is simply used for naming the save file}
}
\value{
An .rdata file containing the rhat values for all species 
     for all years for the region in question.
}
\description{
This function extracts the rhat values for the annual 
occupancy estimates for a set of species for a given region. The 
function loops through Bugs output files within a given location.
}
