% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampBind.R
\name{SampBind}
\alias{SampBind}
\title{SampBind - Combine the multiple outputs from SampPost}
\usage{
SampBind(indata = "../data/model_runs/", innames = NULL,
  output_path = "../data/bound_samples/", group_name = "")
}
\arguments{
\item{indata}{The file path to a location containing .rdata files
of the \code{SampPost()} outputs, one for each species or species group.}

\item{innames}{A vector of file names to be used as the input. If a vector is
not supplied then all .rdata files in the specified folder will be used.}

\item{output_path}{The location where the output will be saved.}

\item{group_name}{The name of the species group we are running, used for
naming output files.}
}
\value{
An .rdata file containing the inputted posterior samples combined into
        a single dataframe.
}
\description{
This function combines sets of samples created by \code{SampPost()}.
             This function can cobine samples with different start and end years
             if max and min years are available in the input attributes.
}
