% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampPost.R
\name{SampPost}
\alias{SampPost}
\title{SampPost - Sample posterior function}
\usage{
SampPost(indata = "../data/model_runs/",
  output_path = "../data/sampled_posterior_1000/", REGION_IN_Q = "psi.fs",
  sample_n = 1000, combined_output = TRUE)
}
\arguments{
\item{indata}{The file path to a location containing .rdata files
of the bugs outputs, one for each species}

\item{output_path}{The location where the output will be saved.}

\item{REGION_IN_Q}{The region for which occupancy estimates will be 
extracted. UK based occupancy model examples include,
 'psi.fs', 'psi.fs.r_GB', 'psi.fs.r_ENGLAND'.}

\item{sample_n}{The number of values extracted from the posterior.}

\item{combined_output}{This specifies whether the output should be a 
single .csv, containing all species (TRUE) or a single .csv file 
per species (FALSE). Default is TRUE.}
}
\value{
A .csv file for each species containing annual occupancy 
      estiamtes for each year as columns, and an iteration and species 
      name column.
}
\description{
This function is used to extract a number of annual 
occupancy estimates from any given posterior. The function loops 
through Bugs output files within a given location.
}
